+++ dubbo-demo/rpc-simple-demo/src/test/java/com/noob/rpc/simple/demo/RpcSimpleDemoApplicationTests.java
@@ -1,4 +1,4 @@
-package com.noob.rpcsimpledemo;
+package com.noob.rpc.simple.demo;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
